Support Acl update

+++ deleteAll.js
@@ -4,26 +4,33 @@ module.exports = function () {
   return new Promise((res, rej) => {
     me
       .count()
+      .limit(1)
       .fetchAll()
       .then((results) => {
         const promises = [];
         count = results.count;
-        for (let i = 0; i < results.results.length; i++) {
-          const item = results.results[i];
+        if (count === 0) {
+          return res();
+        }
+        for (let i = 0; i < results.length; i++) {
+          const item = results[i];
+          if (item === 'count') {
+            continue;
+          }
           promises.push(item.delete());
         }
-        Promise
+        return Promise
           .all(promises)
-          .then((deleteData) => {
-            if (count > 1000) {
-              me.deleteAll();
-            } else {
-              res();
-            }
-          })
-          .catch((err) => {
-            rej(err);
-          });
+      })
+      .then((deleteData) => {
+        if (count > 1000) {
+          me.deleteAll();
+        } else {
+          res();
+        }
+      })
+      .catch((err) => {
+        rej(err);
       });
   });
 };
